requirements.txt
fastapi==0.95.2
gunicorn==20.1.0
load-dotenv==0.1.0
uvicorn==0.22.0
